+++ shelfsense/src/util/ComponentStatusLabel.jsx
