+++ shelfsense/src/util/services/Products/deleteProduct.js
@@ -1,4 +1,4 @@
-import {destroyStoresAndLogout} from "../../destroyStoresAndLogout.js";
+import {destroyStoresAndLogout} from "../Users/destroyStoresAndLogout.js";
 
 export const deleteProduct = async (product) => {
     const BASE_URL = `${import.meta.env.VITE_API_URL}/products`;
