+++ shelfsense/src/util/services/Users/validateSession.js
@@ -1,0 +1,5 @@
+import {getRequest} from "../GetRequestService.jsx";
+
+export const validateSession = async () => {
+    return await getRequest(`session`)
+}
