+++ shelfsense/src/util/services/components/createComponent.js
@@ -1,4 +1,4 @@
-import {destroyStoresAndLogout} from "../../destroyStoresAndLogout.js";
+import {destroyStoresAndLogout} from "../Users/destroyStoresAndLogout.js";
 
 
 export const createComponent = async (componentData) => {
